test: add tests for huge (~5 MiB) files
