+++ .github/workflows/elixir.yml
@@ -14,6 +14,7 @@ jobs:
       matrix:
         otp_version: ['24.0', '23.3', '22.3']
         elixir_version: ['1.11.4', '1.12.1']
+        suite: ['', '--only huge']
     steps:
     - uses: actions/checkout@v2
     - name: Set up Elixir
@@ -32,4 +33,4 @@ jobs:
     - name: Install dependencies
       run: mix do deps.get, deps.compile
     - name: Run tests
-      run: mix test
+      run: mix test ${{ matrix.suite }}
